+++ modules/waf/examples/main.tf
@@ -1,0 +1,4 @@
+module "test_waf" {
+  source            = "../"
+  waf_configuration = var.waf_configuration
+}